Atualiza os testes para utilizar selenium 4.x
Adiciona container selenium no docker-compose

Adiciona um novo teste, test_functional_register.py

Pin na dependencia do selenium no requirements.txt
