+++ .gitignore
@@ -2,3 +2,4 @@
 .env/
 __pycache__/
 .coverage
+mariadb_data
