+++ Dockerfile
@@ -1,10 +1,10 @@
 FROM python:3.9.10-slim-bullseye
 
+RUN apt update; apt install -y python3-mysqldb libmariadb-dev gcc
+
 WORKDIR courseCatalog/
 
 ADD . /courseCatalog/
-
-RUN apt update; apt install -y python3-mysqldb libmariadb-dev gcc
 
 RUN pip install -r requirements.txt
