+++ project/templates/register.html
@@ -52,7 +52,7 @@
           </div>
           {% endif %}
         </div>
-        <button class="btn btn-lg btn-primary" value="Register" type="submit"><i class="fa fa-sign-in" aria-hidden="true"></i> Registry</button>
+        <button id="register" class="btn btn-lg btn-primary" value="Register" type="submit"><i class="fa fa-sign-in" aria-hidden="true"></i> Registry</button>
     </form>
 
   </div>
