+++ requirements.txt
@@ -19,6 +19,6 @@ Werkzeug==1.0.1
 WTForms==2.3.3
 nose
 coverage
-selenium
+selenium==4.21.0
 Rust
 mysqlclient
